+++ src/Resources/app/administration/src/module/configuration/component/update-field-roles/index.js
@@ -1,0 +1,31 @@
+import template from './update-field-roles.html.twig';
+
+const {Component, Mixin} = Shopware;
+
+Component.register('update-field-roles', {
+        template,
+        inject: ['fieldRolesService',],
+        mixins: [
+            Mixin.getByName('notification'),
+            Mixin.getByName('sw-inline-snippet'),
+        ],
+        data() {
+            return {
+                isLoading: false,
+                isSaveSuccessful: false,
+            };
+        },
+
+        methods: {
+            async onClick() {
+                this.isLoading = true;
+                const response = await this.fieldRolesService.sendUpdateFieldRoles();
+                this.isSaveSuccessful = true;
+                this.isLoading = false;
+                this.createNotificationSuccess({
+                    message: this.$tc('configuration.updateFieldRoles.update')
+                })
+            },
+        },
+    },
+);